Fix bug preventing the winner to be determined when all fields are used
This commit fixes a bug that prevented the winner to be determined / the game ending
when the players have used all the fields in the game. This was caused by the
TicTacToeGame#evaluateGameState-method simply returning true when the size of
moves was 9 without evaluating the winner.
Also the descriptions of errors have been improved.